Move Factory methods inside of related type
